+++ controlalgorithm/persistent_data.json
@@ -1,0 +1,5 @@
+{
+    "lat": 53.535411,
+    "lon": -113.507996,
+    "timezone_adjustment": -7
+}
